+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 68.0.0-SNAPSHOT - unreleased
 
+### ⚙️ Technical
+
+* Updated Admin Console's Cluster tab to refresh more frequently.
+
 ## 67.0.0 - 2024-09-03
 
 ### 💥 Breaking Changes (upgrade difficulty: 🟢 LOW - Hoist Core update only)
@@ -41,7 +45,6 @@
 * Fixed inline `SelectEditor` to ensure new value is flushed before grid editing stops.
 * `WebSocketService` now attempts to establish a new connection when app's server instance changes.
 
-
 ### ✨ Styles
 
 * Added CSS variables to support customization of `Badge` component styling.
@@ -49,7 +52,6 @@
 ### 📚 Libraries
 
 * short-unique-id `added @ 5.2`
-* hoist-core 21.0`
 
 ## 66.1.1 - 2024-08-01